Convert test to JUnit 4. Use TesterUtil.addRandomDomain() to allow several tests to run in parallel, backport from Tomcat 7.
This test fails in Tomcat 7, Tomcat 9 as well and is excluded when running tests via build.xml.

This test creates random messages, half of those are failing at receiver side by an explicit "throw new IllegalStateException()".
This causes send() call in the test to fail with a ChannelException, which is uncaught and fails the whole test run.

git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc6.0.x/trunk@1723117 13f79535-47bb-0310-9956-ffa450edef68