Remove redundant copy of catalina.properties from o.a.c.startup. Generate this copy during the ant "compile" task.
In Tomcat 7 this feature was implemented in r1138825 + r1351135.
Note that nowadays the copied file is not included in src archives.

"compile" target needs to depend on "build-prepare"
Otherwise calling this target with an empty output directory fails at javac task with
build.xml: destination directory "<...>\classes" does not exist or is not a directory

git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc6.0.x/trunk@1723827 13f79535-47bb-0310-9956-ffa450edef68